fixed array init and int replaced by size_t

+++ include/task3.h
@@ -1,5 +1,5 @@
-template<class T, int size>
+template<class T, size_t size>
 void map(T arr[], T(*change)(T))
 {
-	for (int i = 0; i < size; i++)arr[i] = change(arr[i]);
+	for (size_t i = 0; i < size; i++)arr[i] = change(arr[i]);
 }
